Updated for respiratory peak finding
Respiration is a piece of junk and under conscious control, so variability in
waveforms is off-the-charts as compared to pulse/ECG. Because of this, we need
to make some tweaks to the peak detection software to improve detection
accuracy. Namely:

gen_thresh()
(1) Reduce the stdev for thresholds from 2.5*std to 1.5*std
(2) Reduce the weightings

rtp_finder()
(1) Implement window-3, first order savgol filter on data
(2) Increase "acceptable" lag in peak finding to 2 datapoints (instead of 1)

get_baseline()
(1) Implement lowpass filter of data before peak finding to improve accuracy

With these, tests seem to indicate between a 75 and 95% accuracy rating;
obviously, a more consistent >85% rating would be ideal, but for now this
should do.